+++ tests/meta/conftest.py
@@ -1,7 +1,6 @@
 """Fixtures for the meta tests"""
 
 from datetime import datetime
-from pathlib import Path
 
 import pytest
 from _pytest import tmpdir
@@ -10,7 +9,7 @@ from cg.store import Store
 from cg.apps.balsamic.fastq import FastqHandler as BalsamicFastqHandler
 from cg.apps.microsalt.fastq import FastqHandler as MicrosaltFastqHandler
 from cg.apps.hk import HousekeeperAPI
-from cg.apps.scoutapi import ScoutAPI
+from cg.apps.scout.scoutapi import ScoutAPI
 from cg.meta.workflow.mip import MipAnalysisAPI
 
 from tests.store_helpers import StoreHelpers
@@ -109,7 +108,6 @@ class MockTB:
             "analysis_sex": {"ADM1": "female", "ADM2": "female", "ADM3": "female"},
             "family": "yellowhog",
             "duplicates": {"ADM1": 13.525, "ADM2": 12.525, "ADM3": 14.525},
-            "genome_build": "hg19",
             "mapped_reads": {"ADM1": 98.8, "ADM2": 99.8, "ADM3": 97.8},
             "mip_version": "v4.0.20",
             "sample_ids": ["2018-20203", "2018-20204"],